Move everything in the module to ease imports
