+++ sequence/config_generators/config_generator_ring.py
