+++ sequence/config_generators/config_generator_star.py
